Adding first draft of snapshot balance data
This is not a perfect list as some of the Dark Coin balances are reflected incorrectly due to problems with rewinding past inner transactions. However, these wallets can be reconciled individually if needed. We are sharing with the community to confirm if the LP balance or Algo balances appear correct. The data is organized as such:

<user wallet address> <dark coin balance> <liquidity token balance (shown as atomic unit)> <algorand extracted from pool(shown as atomic unit)>